支持字节码多字节参数，兼容ExtendedArg机制
本次提交引入 ExtendedArg 指令，实现字节码参数多字节拼接，突破单字节（0~255）限制。重构了字节码生成器和虚拟机执行器，所有参数相关逻辑均支持多字节，兼容 Python 字节码的 ExtendedArg 机制。

+++ PySharp/Compilation/Bytecodes/Instruction.cs
@@ -6,14 +6,14 @@ namespace PySharp.Compilation.Bytecodes;
 internal readonly struct Instruction
 {
     public readonly OpCode OpCode;
-    public readonly int Arg;
+    public readonly byte Arg;
 
     public Instruction(OpCode opCode)
     {
         OpCode = opCode;
         Arg = 0;
     }
-    public Instruction(OpCode opCode, int arg)
+    public Instruction(OpCode opCode, byte arg)
     {
         OpCode = opCode;
         Arg = arg;
